refactor: 상수 파일 game, message로 분리
