+++ src/BridgeGame.js
@@ -1,4 +1,4 @@
-const { COMMAND } = require("./utils/constans");
+const { COMMAND } = require("./constants/game");
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
