+++ src/InputView.js
@@ -1,6 +1,6 @@
 const { Console } = require("@woowacourse/mission-utils");
+const { INPUT_MESSAGE } = require("./constants/message");
 const Validation = require("./Validation");
-const { INPUT_MESSAGE } = require("./utils/constans");
 
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
