+++ src/OutputView.js
@@ -1,5 +1,5 @@
 const { Console } = require("@woowacourse/mission-utils");
-const { OUTPUT_MESSAGE } = require("./utils/constans");
+const { OUTPUT_MESSAGE } = require("./constants/message");
 
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
